+++ src/OpenTelemetry.Instrumentation.AspNet.TelemetryHttpModule/CHANGELOG.md
@@ -5,6 +5,15 @@
 * Updated OpenTelemetry core component version(s) to `1.13.0`.
   ([#3158](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3158))
 
+* **Breaking Change**: This module is no longer responsible for creating activities.
+  The contract of `TelemetryHttpModuleOptions.OnRequestStartedCallback` was changed
+  to `Func<HttpContextBase, ActivityContext, Activity?>?`. The consumer is now
+  responsible for providing function returning `Activity`.
+  ([#3151](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3151))
+
+* Fixed an inaccurate span end of time.
+  ([#3171](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3171))
+
 ## 1.12.0-beta.2
 
 Released 2025-Sep-18